refactor: вынести обновление и удаление документа в методы ExtendedDocument
